+++ crates/engine_types/src/resource.rs
@@ -261,7 +261,7 @@ impl Resource {
     }
 
     pub fn token_symbol(&self) -> Option<&str> {
-        self.metadata.get(TOKEN_SYMBOL)
+        self.metadata.get_str(TOKEN_SYMBOL)
     }
 
     pub fn divisibility(&self) -> u8 {