+++ crates/template_lib/src/resource/builder/fungible.rs
@@ -335,8 +335,8 @@ impl FungibleResourceBuilder {
     ///    .add_metadata("CharacterLvl", "98")
     /// .build();
     /// ```
-    pub fn add_metadata<K: Into<String>, V: Into<String>>(mut self, key: K, value: V) -> Self {
-        self.metadata.insert(key, value);
+    pub fn add_metadata<K: Into<String>, V: tari_bor::Encode<()>>(mut self, key: K, value: V) -> Self {
+        self.metadata.insert(key, &value);
         self
     }
 
@@ -358,7 +358,7 @@ impl FungibleResourceBuilder {
     ///    .metadata("CharacterLvl", "99")
     /// .build();
     /// ```
-    pub fn metadata<K: Into<String>, V: Into<String>>(self, key: K, value: V) -> Self {
+    pub fn metadata<K: Into<String>, V: tari_bor::Encode<()>>(self, key: K, value: V) -> Self {
         self.add_metadata(key, value)
     }
 
@@ -536,7 +536,7 @@ impl FungibleResourceBuilder {
 
     fn build_internal(mut self, mint_arg: Option<MintArg>) -> (ResourceAddress, Option<Bucket>) {
         if let Some(symbol) = self.token_symbol {
-            self.metadata.insert(TOKEN_SYMBOL, symbol);
+            self.metadata.insert(TOKEN_SYMBOL, &symbol);
         }
         ResourceManager::create(
             ResourceType::Fungible,