+++ crates/template_lib/src/resource/builder/non_fungible.rs
@@ -153,8 +153,8 @@ impl NonFungibleResourceBuilder {
     }
 
     /// Adds a new metadata entry to the resource
-    pub fn add_metadata<K: Into<String>, V: Into<String>>(mut self, key: K, value: V) -> Self {
-        self.metadata.insert(key, value);
+    pub fn add_metadata<K: Into<String>, V: tari_bor::Encode<()>>(mut self, key: K, value: V) -> Self {
+        self.metadata.insert(key, &value);
         self
     }
 
@@ -176,7 +176,7 @@ impl NonFungibleResourceBuilder {
     ///    .metadata("CharacterLvl", "99")
     /// .build();
     /// ```
-    pub fn metadata<K: Into<String>, V: Into<String>>(self, key: K, value: V) -> Self {
+    pub fn metadata<K: Into<String>, V: tari_bor::Encode<()>>(self, key: K, value: V) -> Self {
         self.add_metadata(key, value)
     }
 
@@ -310,7 +310,7 @@ impl NonFungibleResourceBuilder {
 
     fn build_internal(mut self, mint_arg: Option<MintArg>) -> (ResourceAddress, Option<Bucket>) {
         if let Some(symbol) = self.token_symbol {
-            self.metadata.insert(TOKEN_SYMBOL, symbol);
+            self.metadata.insert(TOKEN_SYMBOL, &symbol);
         }
 
         ResourceManager::create(